Remove explicit dependency calls from remake
This is because since we have included the file paths in the command call, these files are already classed as explicit dependencies. If we called the files within the body of the function rather than as an argument within the function call then we’d have to explicitly list each file in a depends argument to the target object.

Signed-off-by: egouldo <elisebgould@gmail.com>