Regenerate management lookup with +var mngmnt_unit
This is to support functionality within the generate_case_file_data_frame() fn whereby multiple transects per management unit are aggregated. The ‘raw_data_2014’ file has already been aggregated for the management units with multiple transects. But adding this functionality to the function allows for other users and possibly the 2015 data to be aggregated where there are large management units with multiple transects.

Also note that I have changed the numbering on the field_site_management lookup file to account for the method of aggregation in the generate_case_file_data_frame() fn (group by management_unit).

Signed-off-by: egouldo <elisebgould@gmail.com>