Add Rnetica to packrat.lock w local source install
This is a manual hack to overcome issues with installing packages from source outside of CRAN into a local packrat library.

The DESCRIPTION file for RNETICA package within the ./packrat/lib folder is overwriten with a local URL path pointing to the location of the install source’s filepath, while the `source:` of the file is set to `source` (within DESCRIPTION).

Next you run `packrat::snapshot()` and this updates the lockfile with the correct information.

Signed-off-by: egouldo <elisebgould@gmail.com>