ensure licensekey loaded else fails
This is because the free version of Netica can only handle networks of a limited number of nodes, but the number of nodes in our network is larger than that. During the remake::make() call, the Netica environment link to R is lost between commands, and the function fails because the password is unloaded.

Signed-off-by: egouldo <elisebgould@gmail.com>