add packaged data to remake workflow
packaged data is all raw and summarised data object in format `.rda` to be shipped with the R package inside the folder `./data/`. This is a special directory recognised when the package is built and allows for these objects to be exported inside the package for use by package users.

Signed-off-by: egouldo <elisebgould@gmail.com>